refactor: extract duplicated error handling and magic numbers
**Code Quality Improvements:**

1. **Eliminated Code Duplication** (-45 lines)
   - Extracted 3 identical error handling blocks into reusable utility
   - Created lib/error-handlers.ts with handleJobCreationError()
   - Files cleaned: generation-modal.tsx, empty-states.tsx (2x)

2. **Extracted Magic Numbers to Named Constants**
   - Toast durations: Added TOAST_DURATION.SUCCESS (4000ms) and ERROR (5000ms)
   - Error truncation: MAX_ERROR_SUMMARY_LENGTH (80), TRUNCATED_ERROR_LENGTH (77)
   - Job query limit: RECENT_JOBS_DISPLAY_LIMIT (20)

3. **Improved Error Logging Consistency**
   - Added NODE_ENV guards to generation-task-card.tsx console.error calls
   - Now consistent with rest of codebase (development-only logging)

**Impact:**
- More maintainable: Single source of truth for error handling
- More testable: Utility functions easier to test in isolation
- Better DRY: 45 fewer lines of duplicated code
- Clearer intent: Named constants vs magic numbers

**Verification:**
- ✅ TypeScript: Clean compilation
- ✅ Tests: 397/397 passing
- ✅ Linter: Passing
- ✅ Build: Production build successful
